webapp/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
bleach==3.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
comtypes==1.1.10
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.2
dlib==19.22.0
entrypoints==0.3
enum34==1.1.10
face-recognition==1.4.0
face-recognition-models==0.3.0
facerecog==0.0.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
HandTracking==0.3
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mediapipe==0.8.3.1
mistune==0.8.4
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.15.8
psutil==5.8.0
psycopg2==2.8.6
py==1.10.0
PyAutoGUI==0.9.52
pycaw==20181226
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.7.4
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.5
pywin32==300
pywinpty==0.5.7
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.12.1
schedule==0.6.0
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.4.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
word2number==1.1
xlrd==2.0.1
